refactor: Made `setup` a subcommand of `app`
