+++ src/main/java/dev/jbang/cli/Jbang.java
@@ -46,7 +46,7 @@ import picocli.CommandLine.Model.UsageMessageSpec;
 
 		"" }, versionProvider = VersionProvider.class, subcommands = {
 				Run.class, Build.class, Edit.class, Init.class, Alias.class, Catalog.class, Trust.class, Cache.class,
-				Completion.class, Jdk.class, Version.class, Wrapper.class, Info.class, Setup.class, App.class })
+				Completion.class, Jdk.class, Version.class, Wrapper.class, Info.class, App.class })
 public class Jbang extends BaseCommand {
 
 	@CommandLine.ArgGroup(exclusive = true)
@@ -126,7 +126,7 @@ public class Jbang extends BaseCommand {
 		sections.put("Essentials", asList("run", "build"));
 		sections.put("Editing", asList("init", "edit"));
 		sections.put("Caching", asList("cache", "jdk"));
-		sections.put("Configuration", asList("trust", "alias", "catalog", "app", "setup"));
+		sections.put("Configuration", asList("trust", "alias", "catalog", "app"));
 		sections.put("Other", asList("completion", "info", "version", "wrapper"));
 		CommandGroupRenderer renderer = new CommandGroupRenderer(sections);
 		return renderer;
